requirements.txt
torch>=2.1
numpy
pandas
pyarrow
scikit-learn
pyyaml
matplotlib
tqdm
